Move createPayPalMethod to custom hook
